+++ src/widgets/image_widget.rs
@@ -16,7 +16,9 @@
 use crate::render::callbacks::CallbackRegistry;
 use crate::render::widget::*;
 use crate::render::widget_cache::WidgetContainer;
-use crate::render::widget_config::{WidgetConfig, CONFIG_COLOR_BASE, CONFIG_SIZE};
+use crate::render::widget_config::{
+    CompassPosition, Config, WidgetConfig, CONFIG_COLOR_BASE, CONFIG_IMAGE_POSITION, CONFIG_SIZE,
+};
 use crate::render::Points;
 
 use sdl2::image::LoadTexture;
@@ -65,7 +67,6 @@ pub struct ImageWidget {
     system_properties: HashMap<i32, String>,
     callback_registry: CallbackRegistry,
     image_name: String,
-    image_position: ImagePosition,
     scaled: bool,
 }
 
@@ -80,21 +81,12 @@ impl ImageWidget {
     /// `ImagePosition` will be ignored.  Likewise, if set to `false`, the image will be displayed for
     /// the size of the image, and will be placed in the bounds of the `Widget` based on the position
     /// specified in the `ImagePosition`.
-    pub fn new(
-        image_name: String,
-        image_position: ImagePosition,
-        x: i32,
-        y: i32,
-        w: u32,
-        h: u32,
-        scaled: bool,
-    ) -> Self {
+    pub fn new(image_name: String, x: i32, y: i32, w: u32, h: u32, scaled: bool) -> Self {
         Self {
             config: WidgetConfig::new(x, y, w, h),
             system_properties: HashMap::new(),
             callback_registry: CallbackRegistry::new(),
             image_name,
-            image_position,
             scaled,
         }
     }
@@ -117,26 +109,30 @@ impl Widget for ImageWidget {
         let widget_h = self.get_size(CONFIG_SIZE)[1] as i32;
         let TextureQuery { width, height, .. } = texture.query();
 
-        let texture_x = match self.image_position {
-            ImagePosition::NW | ImagePosition::W | ImagePosition::SW => self.get_config().to_x(0),
+        let texture_x = match self.get_compass(CONFIG_IMAGE_POSITION) {
+            CompassPosition::NW | CompassPosition::W | CompassPosition::SW => {
+                self.get_config().to_x(0)
+            }
 
-            ImagePosition::N | ImagePosition::Center | ImagePosition::S => {
+            CompassPosition::N | CompassPosition::Center | CompassPosition::S => {
                 self.get_config().to_x((widget_w - width as i32) / 2)
             }
 
-            ImagePosition::NE | ImagePosition::E | ImagePosition::SE => {
+            CompassPosition::NE | CompassPosition::E | CompassPosition::SE => {
                 self.get_config().to_x(widget_w - width as i32)
             }
         };
 
-        let texture_y = match self.image_position {
-            ImagePosition::NW | ImagePosition::N | ImagePosition::NE => self.get_config().to_y(0),
+        let texture_y = match self.get_compass(CONFIG_IMAGE_POSITION) {
+            CompassPosition::NW | CompassPosition::N | CompassPosition::NE => {
+                self.get_config().to_y(0)
+            }
 
-            ImagePosition::W | ImagePosition::Center | ImagePosition::E => {
+            CompassPosition::W | CompassPosition::Center | CompassPosition::E => {
                 self.get_config().to_y((widget_h - height as i32) / 2)
             }
 
-            ImagePosition::SW | ImagePosition::S | ImagePosition::SE => {
+            CompassPosition::SW | CompassPosition::S | CompassPosition::SE => {
                 self.get_config().to_y(widget_h - height as i32)
             }
         };
@@ -160,6 +156,13 @@ impl Widget for ImageWidget {
                 ),
             )
             .unwrap();
+        }
+    }
+
+    /// Responds to a screen redraw only if the `CONFIG_IMAGE_POSITION` key was changed.
+    fn on_config_changed(&mut self, _k: u8, _v: Config) {
+        if _k == CONFIG_IMAGE_POSITION {
+            self.get_config().set_invalidate(true);
         }
     }